requirements.txt
affine==2.4.0
attrs==25.4.0
certifi==2026.1.4
click==8.3.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
geopandas==1.1.2
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.5.0
shapely==2.1.2
six==1.17.0
tzdata==2025.3
libpysal
esda
